fix: first connect home actuve
